accept deprecated kubernetes.io/ingress.class annotation
